requirements.txt
aiofile==3.9.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
aiosqlite3==0.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
asyncer==0.0.17
attrs==25.4.0
Authlib==1.6.9
beartype==0.22.9
bidict==0.23.1
cachetools==7.0.3
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
chainlit==2.10.0
charset-normalizer==3.4.5
chevron==0.14.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
cuid==0.4
cyclopts==4.7.0
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
fastmcp==3.1.0
filelock==3.25.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.11
langchain-core==1.2.18
langchain-mcp-adapters==0.2.2
langchain-openai==1.1.11
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
Lazify==0.4.0
literalai==0.1.201
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
openai==2.26.0
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-agno==0.53.0
opentelemetry-instrumentation-alephalpha==0.53.0
opentelemetry-instrumentation-anthropic==0.53.0
opentelemetry-instrumentation-bedrock==0.53.0
opentelemetry-instrumentation-chromadb==0.53.0
opentelemetry-instrumentation-cohere==0.53.0
opentelemetry-instrumentation-crewai==0.53.0
opentelemetry-instrumentation-google-generativeai==0.53.0
opentelemetry-instrumentation-groq==0.53.0
opentelemetry-instrumentation-haystack==0.53.0
opentelemetry-instrumentation-lancedb==0.53.0
opentelemetry-instrumentation-langchain==0.53.0
opentelemetry-instrumentation-llamaindex==0.53.0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-marqo==0.53.0
opentelemetry-instrumentation-mcp==0.53.0
opentelemetry-instrumentation-milvus==0.53.0
opentelemetry-instrumentation-mistralai==0.53.0
opentelemetry-instrumentation-ollama==0.53.0
opentelemetry-instrumentation-openai==0.53.0
opentelemetry-instrumentation-openai-agents==0.53.0
opentelemetry-instrumentation-pinecone==0.53.0
opentelemetry-instrumentation-qdrant==0.53.0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-replicate==0.53.0
opentelemetry-instrumentation-requests==0.61b0
opentelemetry-instrumentation-sagemaker==0.53.0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-threading==0.61b0
opentelemetry-instrumentation-together==0.53.0
opentelemetry-instrumentation-transformers==0.53.0
opentelemetry-instrumentation-urllib3==0.61b0
opentelemetry-instrumentation-vertexai==0.53.0
opentelemetry-instrumentation-voyageai==0.53.0
opentelemetry-instrumentation-watsonx==0.53.0
opentelemetry-instrumentation-weaviate==0.53.0
opentelemetry-instrumentation-writer==0.53.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-semantic-conventions-ai==0.4.15
opentelemetry-util-http==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pathable==0.5.0
platformdirs==4.9.4
propcache==0.4.1
protobuf==6.33.5
py-key-value-aio==0.4.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyperclip==1.11.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlite-vec==0.1.6
sse-starlette==3.3.2
starlette==0.52.1
syncer==2.0.3
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tqdm==4.67.3
traceloop-sdk==0.53.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
